Swapped x and y with a and b in Section 2.
Previously [1], Ion pointed out that it's a bit confusing that we use x
and y to range over commands while our commands are of the form `x = 1`
and `y = 1`. Ion modified section 2 to let a and b range over commands.
This commit flips the decision, keeping x and y ranging over commands,
but modifying Figure 1 to have commands of the form `a = 1` and `b = 2`.
This keeps section 2 more consistent with the rest of the paper, where
we let x and y range over commands.

[1]: 8cf056fe018a6d89e4ffca5fd8c18f4c767de013