Progress on background section.
This paper is tricky because there is a lot of background material you
need to understand in order to understand some of the more complex
protocols later on. JSys doesn't have a page limit, which means we can
explain everything as deeply as we want, but I also don't want the paper
to be so long that a reader finds it intimidating.